Update for region data source move to NAS
