+++ internal/provider/provider.go
@@ -543,8 +543,8 @@ func Provider() *schema.Provider {
 			"aws_redshift_cluster":                           redshift.DataSourceCluster(),
 			"aws_redshift_orderable_cluster":                 redshift.DataSourceOrderableCluster(),
 			"aws_redshift_service_account":                   redshift.DataSourceServiceAccount(),
-			"aws_region":                                     ec2.DataSourceRegion(),
-			"aws_regions":                                    ec2.DataSourceRegions(),
+			"aws_region":                                     nas.DataSourceRegion(),
+			"aws_regions":                                    nas.DataSourceRegions(),
 			"aws_resourcegroupstaggingapi_resources":         resourcegroupstagging.DataSourceResources(),
 			"aws_route":                                      ec2.DataSourceRoute(),
 			"aws_route_table":                                ec2.DataSourceRouteTable(),
