+++ internal/service/nas/region_data_source.go
@@ -1,4 +1,4 @@
-package ec2
+package nas
 
 import (
 	"fmt"
