+++ internal/service/nas/regions_data_source.go
@@ -1,4 +1,4 @@
-package ec2
+package nas
 
 import (
 	"fmt"
@@ -8,6 +8,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+	tfec2 "github.com/hashicorp/terraform-provider-aws/internal/service/ec2"
 )
 
 func DataSourceRegions() *schema.Resource {
@@ -15,7 +16,7 @@ func DataSourceRegions() *schema.Resource {
 		Read: dataSourceRegionsRead,
 
 		Schema: map[string]*schema.Schema{
-			"filter": dataSourceFiltersSchema(),
+			"filter": tfec2.DataSourceFiltersSchema(),
 			"names": {
 				Type:     schema.TypeSet,
 				Computed: true,
@@ -36,7 +37,7 @@ func dataSourceRegionsRead(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("[DEBUG] Reading regions.")
 	request := &ec2.DescribeRegionsInput{}
 	if v, ok := d.GetOk("filter"); ok {
-		request.Filters = buildFiltersDataSource(v.(*schema.Set))
+		request.Filters = tfec2.BuildFiltersDataSource(v.(*schema.Set))
 	}
 	if v, ok := d.GetOk("all_regions"); ok {
 		request.AllRegions = aws.Bool(v.(bool))
